Perform orientation and magnitude tests

+++ src/optical_flow/vector_statistics.h
@@ -11,7 +11,7 @@
 #ifndef SRC_OPTICAL_FLOW_VECTOR_STATISTICS_H_
 #define SRC_OPTICAL_FLOW_VECTOR_STATISTICS_H_
 namespace oflow {
-using vector_type = std::array<std::vector<cv::Point2f>, 2>;
+using vector_type = std::array<std::vector<cv::Point2d>, 2>;
 class VectorStatistics {
  public:
   explicit VectorStatistics(const vector_type &points) : points_{points} {};
